requirements/base.txt
Django==3.0.10
django-allauth==0.42.0
django-envelope==1.4
django-environ==0.4.5
django-filter==2.3.0
django-octicons==1.0.2
django-scopes==1.2.0
django-stronghold==0.4.0
matplotlib==3.3.2
postgres==3.0.0
python-dateutil==2.8.1
pytz==2020.1
whitenoise==5.0.1
